+++ src/components/navigation/Navbar.stories.tsx
@@ -7,6 +7,7 @@ import logoImageLight from "../../public/generic/logo-light.svg";
 import { ColourSchemeButton } from "../controls/ColourSchemeButton";
 import { User } from "../controls/User";
 import { MockLink } from "../../utils/MockLink";
+import { Logo } from "../controls/Logo";
 
 const meta: Meta<typeof Navbar> = {
   title: "SciReactUI/Navigation/Navbar",
@@ -170,6 +171,32 @@ export const WithNonThemeLogo: Story = {
     docs: {
       description: {
         story: "A separate image can also be referenced.",
+      },
+    },
+  },
+};
+
+export const WithThemeLogoAsChild: Story = {
+  args: {
+    children: (
+      <>
+        <NavLinks key="links">
+          <NavLink href="#" key="first">
+            First
+          </NavLink>
+          <NavLink href="#" key="second">
+            Second
+          </NavLink>
+        </NavLinks>
+        <Logo style={{ marginRight: "100px", transform: "rotateX(180deg)" }} />
+      </>
+    ),
+  },
+  parameters: {
+    docs: {
+      description: {
+        story:
+          "You can pass the logo in as a child instead of part of the Navbar for full control.",
       },
     },
   },