+++ changelog.md
@@ -2,11 +2,24 @@ SciReactUI Changelog
 ====================
 
 
+[v0.2.1alpha] - 2025-??-??
+--------------------
+
+### Added
+- Logo component, to easily add the theme logo to anywhere
+
+### Fixed
+- 
+
+### Changed
+- ImageColorSchemeSwitch takes parameter to interchange images on theme mode.
+
 [v0.2.0] - 2025-06-11
 ---------------------
 
 ### Fixed
 - Styles added to Navbar and Footer incorrectly remove built in styles.
+- Logo not appearing when no dark src set in dark mode.
 
 ### Changed
 - Breadcrumbs component takes optional linkComponent prop for page routing.